Unify integer types + check integer overflow
Use usize instead of c_long/u64/u32

+++ sysaug/src/clone.rs
@@ -11,7 +11,7 @@ use tracing::{event, Level};
 lazy_static! {
     static ref SYSCALL_NAMES: HashSet<ptrace::SysNum> = {
         let mut ans = HashSet::new();
-        ans.insert(libc::SYS_clone);
+        ans.insert(libc::SYS_clone as usize);
         ans
     };
 }
